* develop: (152 commits)
  chore(release): prepare for v0.1.7
  docs(book): record the standards text audit
  fix(nmea): reserved sequence id costs binding not data
  fix(nmea): send each pgn at its default priority
  fix(j1939): expose speed prefix decode to python
  docs: cite annex j for auxiliary, not part 11
  fix(tc): set process data priority per command value
  fix(diag): skip unknown functionalities on decode
  fix(fs): ignore reserved bits on volume requests
  feat(tim): type the guidance and speed facilities
  fix(autodrive): zero reserved bits in facility request
  docs(autodrive): correct the TIM and facilities answer
  fix(autodrive): gate driving on tractor facilities
  test(tool): drive the keyboard input path end to end
  fix(tool): survive cramped terminals in every tui
  fix(tool): clip key hints so a small terminal cannot panic
  docs(book): drop the deleted legacy file-transfer types
  refactor: remove the guidance plugin, superseded by autodrive
  refactor(examples): move guidance examples onto autodrive
  docs(book): add a runnable autodrive driving example
  ...

+++ src/isobus/tc/ddi_database/ddi_entry_types.rs
@@ -8,7 +8,14 @@ pub struct DDIDefinition {
     pub unit: &'static str,
     pub resolution: f64,
     pub min_value: i32,
-    pub max_value: i32,
+    /// Widened past `i32` because the condensed DDIs (161-176, 290-305,
+    /// 367-382, 517, 518) have a CANBus range of 0..4294967295 — and once
+    /// unused sections are filled with 11 as DDI 161 requires, 0xFFFFFFFF is
+    /// the *normal* value for any implement with fewer than 16 sections. The
+    /// clamp made the ordinary encoding read as out of range through
+    /// `ddi_display_range` and `ddi_data_dictionary_entry`, the two accessors
+    /// an integrator range-checks against before transmitting.
+    pub max_value: i64,
 }
 
 /// Legacy alias.
